HealthTextAnalytics/HealthTextAnalytics/__init__.py
import logging
import azure.functions as func
import json
from .analyzetxt import txtanalyze



def main(req: func.HttpRequest) -> func.HttpResponse:
    logging.info('Python HTTP trigger function processed a request.')

    name = req.params.get('name')
    output = txtanalyze()

    
    return func.HttpResponse(json.dumps(output))

    if name:
        return func.HttpResponse(f"Hello, {name}. This HTTP triggered function executed successfully.")
    else:
        return func.HttpResponse(
             "This HTTP triggered function executed successfully. Pass a name in the query string or in the request body for a personalized response.",
             status_code=200
        )
