HealthTextAnalytics/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-ai-textanalytics==5.1.0
azure-cognitiveservices-speech==1.18.0
azure-common==1.1.27
azure-core==1.16.0
azure-functions==1.7.2
azure-storage-blob==12.8.1
requests==2.26.0
requests-oauthlib==1.3.0
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
cryptography==3.4.7
debugpy==1.3.0
decorator==5.0.9
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
msrest==0.6.21
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
pyzmq==22.1.0
six==1.16.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5